backend/requirements.txt
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.2

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9

# Database
sqlalchemy==2.0.29

# Live prices
yfinance==0.2.38

# API
fastapi==0.110.0
uvicorn==0.29.0